+++ src/main/kotlin/com/everywaffle/team3server/comment/model/CommentLikeEntity.kt
@@ -1,6 +1,6 @@
-package com.everywaffle.team3server.comment.repository
+package com.everywaffle.team3server.comment.model
 
-import com.everywaffle.team3server.user.repository.UserEntity
+import com.everywaffle.team3server.user.model.UserEntity
 import jakarta.persistence.Entity
 import jakarta.persistence.FetchType
 import jakarta.persistence.GeneratedValue
