+++ src/main/kotlin/com/everywaffle/team3server/comment/repository/CommentRepository.kt
@@ -1,5 +1,6 @@
 package com.everywaffle.team3server.comment.repository
 
+import com.everywaffle.team3server.comment.model.CommentEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface CommentRepository : JpaRepository<CommentEntity, Long>
